separated services into distinct pages and allowed for toggling debug images
